Updating tooling notes and shifting snippet parsing to midprocessor (#592)
* Updating tooling notes and shifting snippet parsing completely to midprocessor

* Including CF-related @types

* Including CF-related @types

* Including CF-related @types
