+++ generate-docs/scripts/preprocessor.ts
@@ -4,7 +4,6 @@ import { fetchAndThrowOnError, DtsBuilder } from './util';
 import { promptFromList } from './simple-prompts';
 import * as path from "path";
 import * as fsx from 'fs-extra';
-import yaml = require('js-yaml');
 
 tryCatch(async () => {
     // ----
@@ -167,62 +166,6 @@ tryCatch(async () => {
 
     console.log("create file: office-runtime.d.ts");
     fsx.writeFileSync('../api-extractor-inputs-office-runtime/office-runtime.d.ts', definitionsForORun);
-
-    // ----
-    // Process Snippets
-    // ----
-    console.log("\nRemoving old snippets input files...");
-
-    const scriptInputsPath = path.resolve("../script-inputs");
-    fsx.readdirSync(scriptInputsPath)
-        .filter(filename => filename.indexOf("snippets") > 0)
-        .forEach(filename => fsx.removeSync(scriptInputsPath + '/' + filename));
-
-    console.log("\nCreating snippets file...");
-
-    console.log("\nReading from: https://raw.githubusercontent.com/OfficeDev/office-js-snippets/master/snippet-extractor-output/snippets.yaml");
-    fsx.writeFileSync("../script-inputs/script-lab-snippets.yaml", await fetchAndThrowOnError("https://raw.githubusercontent.com/OfficeDev/office-js-snippets/master/snippet-extractor-output/snippets.yaml", "text"));
-
-    console.log("\nReading from files: " + path.resolve("../../docs/code-snippets"));
-
-    const snippetsSourcePath = path.resolve("../../docs/code-snippets");
-    let localCodeSnippetsString : string = "";
-    fsx.readdirSync(path.resolve(snippetsSourcePath))
-        .filter(name => name.endsWith('.yaml') || name.endsWith('.yml'))
-        .forEach((filename, index) => {
-            localCodeSnippetsString += fsx.readFileSync(`${snippetsSourcePath}/${filename}`).toString() + "\r\n";
-        });
-
-    fsx.writeFileSync("../script-inputs/local-repo-snippets.yaml", localCodeSnippetsString);
-
-    // Parse the YAML into an object/hash set.
-    let snippets: Object = yaml.load(localCodeSnippetsString);
-
-    // If a duplicate key exists, merge the Script Lab example(s) into the item with the existing key.
-    let scriptLabSnippets: Object = yaml.load(fsx.readFileSync(`../script-inputs/script-lab-snippets.yaml`).toString());
-    for (const key of Object.keys(scriptLabSnippets)) {
-        if (snippets[key]) {
-            console.log("Combining local and Script Lab snippets for: " + key);
-            snippets[key] = snippets[key].concat(scriptLabSnippets[key]);
-        } else {
-            snippets[key] = scriptLabSnippets[key];
-        }
-    }
-
-    console.log("\nWriting snippets to: " + path.resolve("../json/snippets.yaml"));
-
-    fsx.writeFileSync("../json/snippets.yaml", yaml.safeDump(
-        snippets,
-        {sortKeys: <any>((a: string, b: string) => {
-            if (a < b) {
-                return -1;
-            } else if (a > b) {
-                return 1;
-            } else {
-                return 0;
-            }
-        })}
-    ));
 
     console.log("\nPreprocessor script complete!");
